requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
-e git+https://github.com/django/django.git@53d229ff632c2a3e547f2820a94239f38ba4d4ac#egg=Django
django-braces==1.14.0
django-cleanup==4.0.0
django-extra-views==0.13.0
django-filter==2.2.0
django-grappelli==2.14.1
django-heroku==0.3.1
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.9
oauthlib==3.1.0
Pillow==7.1.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==1.5
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.9
whitenoise==5.0.1
